+++ lib/views/screens/home_page/widgets/banner_item.dart
@@ -2,11 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class BannerItem extends StatefulWidget {
-const   BannerItem({Key? key}) : super(key: key);
+  const BannerItem({Key? key}) : super(key: key);
 
   @override
   State<BannerItem> createState() => _BannerItemState();
 }
+
 class _BannerItemState extends State<BannerItem> {
   int selectedIndex = 0;
   @override
@@ -28,56 +29,49 @@ class _BannerItemState extends State<BannerItem> {
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.circular(10.w),
                   image: DecorationImage(
-                    image: AssetImage(bannerList[index].thumnail),
-                    fit: BoxFit.cover
-                    ),
+                      image: AssetImage(bannerList[index].thumnail),
+                      fit: BoxFit.cover),
                 ),
               );
             }),
-            ),
+          ),
         ),
-       SizedBox(height: 8.h,),
-       Row(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          ...List.generate(bannerList.length, (index) => Indicator(
-            isActive: selectedIndex == index ? true : false,
-          )),
-        ],
-       ) ,
+        SizedBox(
+          height: 8.h,
+        ),
+        Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            ...List.generate(
+              bannerList.length,
+              (index) => AnimatedContainer(
+                duration: const Duration(microseconds: 350),
+                margin: EdgeInsets.symmetric(horizontal: 4.w),
+                width: 6.w,
+                height: 6.h,
+                decoration: BoxDecoration(
+                  color: selectedIndex == index
+                      ? const Color(0xffffffff)
+                      : const Color(0xffffffff).withOpacity(0.5),
+                ),
+              ),
+            ),
+          ],
+        ),
       ],
     );
   }
 }
 
-class Indicator extends StatelessWidget {
-  final bool isActive;
-  const Indicator({
-    Key? key,
-   required this.isActive,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedContainer(
-      duration: const Duration(microseconds: 350),
-      margin:  EdgeInsets.symmetric(horizontal: 4.w),
-      width:  6.w,
-      height: 6.h,
-      decoration:  BoxDecoration(
-        color: isActive ?  const Color(0xffffffff) : const Color(0xffffffff).withOpacity(0.5),
-      ),
-    );
-  }
-}
 class AppBanner {
   final int id;
   final String thumnail;
 
   AppBanner(this.id, this.thumnail);
 }
+
 List<AppBanner> bannerList = [
-   AppBanner(1, "assets/images/Group 120.png"),
-   AppBanner(2, "assets/images/Group 120.png"),
-   AppBanner(3, "assets/images/Group 120.png"),
+  AppBanner(1, "assets/images/Group 120.png"),
+  AppBanner(2, "assets/images/Group 120.png"),
+  AppBanner(3, "assets/images/Group 120.png"),
 ];
