+++ lib/views/screens/home_page/widgets/product_home_page/components/endows_partner.dart
@@ -1,0 +1,63 @@
+import 'package:accum_points_app/views/components_app/endow.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class EndowsPartner extends StatelessWidget {
+  const EndowsPartner({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    List<Map<String, dynamic>> brands2 = [
+      {
+        "image": "assets/images/hello summer.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "The coffee house",
+        'numberPoints': 50,
+      },
+      {
+        "image": "assets/images/super sale.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "Gemini",
+        'numberPoints': 200,
+      },
+      {
+        "image": "assets/images/hello summer.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "The coffee house",
+        'numberPoints': 50,
+      },
+      {
+        "image": "assets/images/super sale.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "Gemini",
+        'numberPoints': 200,
+      },
+      {
+        "image": "assets/images/hello summer.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "The coffee house",
+        'numberPoints': 50,
+      },
+      {
+        "image": "assets/images/super sale.png",
+        "title": "Ưu đãi giảm 50.000 đồng",
+        "nameBrand": "Gemini",
+        'numberPoints': 200,
+      },
+    ];
+    return Container(
+      padding: EdgeInsets.symmetric(horizontal: 18.w),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: List.generate(
+          brands2.length,
+          (index) => Endow(
+              image: brands2[index]['image'],
+              titleEndow: brands2[index]['title'],
+              nameBrandEndow: brands2[index]['nameBrand'],
+              numberPoint: brands2[index]['numberPoints']),
+        ),
+      ),
+    );
+  }
+}
